[Add Feature] Hondo Housing System - A site-centric system that allows administrators to easily create player housing anywhere. - Sites contain a number of plots and each plot is home to one structure. Players use a terminal at the site to purchase a structure. - Over 20 pre-defined site layouts (that are easy to customize!). - Contains a building tour system so players can see them first hand before making their purchase. - Enable the "simple" planet and teleport to 1000x 1000y to see a demonstration city that was created using this system. - Full documentation is located in doc/features/add-housingsystem.md
Files:
	modified:   MMOCoreORB/bin/conf/config.lua
	modified:   MMOCoreORB/bin/scripts/object/allobjects.lua
	new file:   MMOCoreORB/bin/scripts/object/tangible/hondo/objects.lua
	modified:   MMOCoreORB/bin/scripts/object/tangible/hondo/serverobjects.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/housing_system.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/housing_system_layouts.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_corellia.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_dantooine.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_dathomir.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_endor.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_lok.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_naboo.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_rori.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_simple.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_simple_demo_city.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_talus.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_tatooine.lua
	new file:   MMOCoreORB/bin/scripts/screenplays/hondo/housing_system/locations_yavin4.lua
	modified:   MMOCoreORB/bin/scripts/screenplays/screenplays.lua
	modified:   MMOCoreORB/src/server/zone/managers/director/DirectorManager.cpp
	modified:   MMOCoreORB/src/server/zone/managers/director/DirectorManager.h
	modified:   MMOCoreORB/src/server/zone/objects/creature/LuaCreatureObject.cpp
	modified:   MMOCoreORB/src/server/zone/objects/creature/LuaCreatureObject.h
	modified:   MMOCoreORB/src/server/zone/objects/creature/commands/hondoCommand.h
	modified:   MMOCoreORB/src/server/zone/objects/player/LuaPlayerObject.cpp
	modified:   MMOCoreORB/src/server/zone/objects/player/LuaPlayerObject.h
	modified:   MMOCoreORB/src/server/zone/objects/player/sessions/DestroyStructureSession.idl
	modified:   MMOCoreORB/src/server/zone/objects/player/sessions/DestroyStructureSessionImplementation.cpp
	modified:   MMOCoreORB/src/server/zone/objects/structure/events/StructureMaintenanceTask.cpp

+++ MMOCoreORB/src/server/zone/objects/structure/events/StructureMaintenanceTask.cpp
@@ -33,6 +33,10 @@ void StructureMaintenanceTask::run() {
 	ManagedReference<CreatureObject*> owner = strongRef->getOwnerCreatureObject();
 
 	if (owner == NULL || !owner->isPlayerCreature()) {
+		// Legend of Hondo - Exception to allow temp buildings for housing system
+		if (owner->getFirstName().toLowerCase() == "r3-0wn3r")
+			return;
+
 		info("Player structure has NULL owner, destroying.", true);
 		StructureManager::instance()->destroyStructure(strongRef);
 		return;
@@ -41,6 +45,10 @@ void StructureMaintenanceTask::run() {
 	ManagedReference<PlayerObject*> ghost = owner->getPlayerObject();
 
 	if (ghost == NULL) {
+		// Legend of Hondo - Exception to allow temp buildings for housing system
+		if (owner->getFirstName().toLowerCase() == "r3-0wn3r")
+			return;
+
 		info("Player structure has NULL owner ghost, destroying.", true);
 
 		StructureManager::instance()->destroyStructure(strongRef);
